Add a deploy script that sets the stack name and calls sam deploy
