+++ buildspec.yaml
@@ -21,4 +21,4 @@ phases:
       - sam build
   post_build:
     commands:
-      sam deploy --parameter-overrides UrlExpiration=$UrlExpiration MaxFileSize=$MaxFileSize --stack-name="basic-ruby-${CODEBUILD_GIT_BRANCH}"
+      - ruby ./scripts/deploy.rb
