+++ image_processing/create_thumbnails.rb
@@ -1,6 +1,7 @@
 module ImageProcessing
   module CreateThumbnails
     def self.handler(event:, context:)
+      p "A change to deploy"
       p event
       p context
     end
